rust_fuzzylogic

Module prelude

Prelude: convenient re-exports for common fuzzy logic items.

Bring this module into scope to get the core trait and types without importing each one individually:

use rust_fuzzylogic::prelude::*;

Note: you still need to expose this module from lib.rs with pub mod prelude; to make it available to users.

Re-exports

pub use crate::error::FuzzyError;

pub use crate::error::Result;

pub use crate::Float;

pub use crate::membership::trapezoidal::Trapezoidal;

pub use crate::membership::MembershipFn;

pub use crate::membership::Gaussian;

pub use crate::membership::Triangular;

pub use crate::ops::FuzzyOps;

pub use crate::term::Term;

pub use crate::sampler::Sampler;

pub use crate::sampler::UniformSampler;